LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-layout-expected.txt

Testing the MacOSInlineMediaControls layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.controlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.controlsBar.children[1] instanceof LayoutNode is true
PASS mediaControls.controlsBar.children[2] instanceof ButtonsContainer is true
PASS mediaControls.timeControl.x is 118
PASS mediaControls.timeControl.width == expectedTimeControlWidth is true
PASS mediaControls.controlsBar.children[0].children is [mediaControls.playPauseButton, mediaControls.skipBackButton]
PASS mediaControls.controlsBar.children[2].children is [mediaControls.muteButton, mediaControls.airplayButton, mediaControls.pipButton, mediaControls.tracksButton, mediaControls.fullscreenButton]
PASS mediaControls.volumeSlider.parent.x == expectedVolumeSliderContainerX is true
PASS successfullyParsed is true

TEST COMPLETE
